divide the list view code into multiple classes
